simplify build logic and add cli/examples
My previous approach to porting the build system was to directly follow the
CMake logic, giving options for all backends. This was a mistake. AFAIK the reason
libgit2 supports so many backends is to give upstream package maintainers
options for their platforms.

Think about it this way: which HTTPS backend should I use? Well, best one of course.
This leads to 2 questions: which backend is the best, and why give an option
if you already know which is best? If you're a package maintainer, the best
backend is whatever follows your platform guidelines, which explains the number of options.
We don't have such requirements, so I've just hard-coded whatever I deemed the
best option at the time. People who need certain options can always upstream a
patch to this fork.

I've also added run steps for the cli and example code.